move encoder functions to nfctag/lf
