+++ firmware/application/src/rfid/nfctag/lf/lf_tag_em.h
@@ -22,5 +22,6 @@ int lf_tag_em410x_data_loadcb(tag_specific_type_t type, tag_data_buffer_t *buffe
 int lf_tag_em410x_data_savecb(tag_specific_type_t type, tag_data_buffer_t *buffer);
 bool lf_tag_em410x_data_factory(uint8_t slot, tag_specific_type_t tag_type);
 uint64_t em410x_id_to_memory64(uint8_t id[5]);
+void em410x_encoder(uint8_t *pData, uint8_t *pOut);
 
 #endif
