+++ firmware/application/src/rfid/reader/lf/lf_viking_data.h
@@ -12,7 +12,6 @@ extern "C"
 #endif
 
 void init_viking_hw(void);
-void viking_encoder(uint8_t *pData, uint8_t *pOut);
 uint8_t viking_decoder(uint8_t *pData, uint8_t size, uint8_t *pOut);
 uint8_t viking_read(uint8_t *uid, uint32_t timeout_ms);
